chore: move code to proper places
