+++ crates/core/generate-pie/src/types/mod.rs
@@ -1,7 +1,13 @@
+pub mod block;
 pub mod chain_config;
 pub mod os_hints;
 pub mod pie;
+pub mod proof;
+pub mod transaction;
 
+pub use block::*;
 pub use chain_config::*;
 pub use os_hints::*;
 pub use pie::*;
+pub use proof::*;
+pub use transaction::*;
